Module epic_core::core

Core types

Re-exports

pub use self::block::*;

pub use self::block_sums::*;

pub use self::committed::Committed;

pub use self::compact_block::*;

pub use self::id::ShortId;

pub use self::transaction::*;

Modules

block

Blocks and blockheaders

block_sums

BlockSums per-block running totals for utxo_sum and kernel_sum. Allows fast “full” verification of kernel sums at a given block height.

committed

The Committed trait and associated errors.

compact_block

Compact Blocks.

foundation

hash

Hash Function

id

short ids for compact blocks

merkle_proof

Merkle Proofs

pmmr

Persistent and prunable Merkle Mountain Range implementation. For a high level description of MMRs, see:

transaction

Transactions

Enums

Error

Common errors

Functions

amount_from_hr_string

Common method for parsing an amount from human-readable, and converting to internally-compatible u64

amount_to_hr_string

Common method for converting an amount to a human-readable string