Struct epic_core::core::block::Block

pub struct Block {
    pub header: BlockHeader,
    // some fields omitted
}

A block as expressed in the MimbleWimble protocol. The reward is non-explicit, assumed to be deducible from block height (similar to bitcoin’s schedule) and expressed as a global transaction fee (added v.H), additive to the total of fees ever collected.

Fields
header: BlockHeader

The header with metadata and commitments to the rest of the data

Implementations

impl Block

pub fn new(
prev: &BlockHeader,
txs: Vec<Transaction>,
difficulty: Difficulty,
reward_output: (Output, TxKernel)
) -> Result<Block, Error>

Builds a new block from the header of the previous block, a vector of transactions and the private key that will receive the reward. Checks that all transactions are valid and calculates the Merkle tree.

TODO - Move this somewhere where only tests will use it. *** Only used in tests. ***

pub fn hydrate_from(
cb: CompactBlock,
txs: Vec<Transaction>
) -> Result<Block, Error>

Hydrate a block from a compact block. Note: caller must validate the block themselves, we do not validate it here.

pub fn with_header(header: BlockHeader) -> Block

Build a new empty block from a specified header

pub fn from_reward(
prev: &BlockHeader,
txs: Vec<Transaction>,
reward_out: Output,
reward_kern: TxKernel,
difficulty: Difficulty
) -> Result<Block, Error>

Builds a new block ready to mine from the header of the previous block, a vector of transactions and the reward information. Checks that all transactions are valid and calculates the Merkle tree.

pub fn from_coinbases(
prev: &BlockHeader,
txs: Vec<Transaction>,
reward: (Output, TxKernel),
foundation: (Output, TxKernel),
difficulty: Difficulty
) -> Result<Block, Error>

Builds a new block ready to mine from the header of the previous block, a vector of transactions and the vector with the reward and foundation coinbase. Checks that all transactions are valid and calculates the Merkle tree.

pub fn with_reward(self, reward_out: Output, reward_kern: TxKernel) -> Block

Consumes this block and returns a new block with the coinbase output and kernels added

pub fn with_coinbase(
self,
reward: (Output, TxKernel),
foundation: (Output, TxKernel)
) -> Block

Consumes this block and returns a new block with the coinbase output and kernels added

pub fn inputs(&self) -> &Vec<Input>

Get inputs

pub fn inputs_mut(&mut self) -> &mut Vec<Input>

Get inputs mutable

pub fn outputs(&self) -> &Vec<Output>

Get outputs

pub fn outputs_mut(&mut self) -> &mut Vec<Output>

Get outputs mutable

pub fn kernels(&self) -> &Vec<TxKernel>

Get kernels

pub fn kernels_mut(&mut self) -> &mut Vec<TxKernel>

Get kernels mut

pub fn total_fees(&self) -> u64

Sum of all fees (inputs less outputs) in the block

pub fn cut_through(self) -> Result<Block, Error>

Matches any output with a potential spending input, eliminating them from the block. Provides a simple way to cut-through the block. The elimination is stable with respect to the order of inputs and outputs. Method consumes the block.

pub fn validate_read(&self) -> Result<(), Error>

“Lightweight” validation that we can perform quickly during read/deserialization. Subset of full validation that skips expensive verification steps, specifically -

rangeproof verification (on the body)

kernel signature verification (on the body)

coinbase sum verification

kernel sum verification

pub fn validate(
&self,
prev_kernel_offset: &BlindingFactor
) -> Result<Commitment, Error>

Validates all the elements in a block that can be checked without additional data. Includes commitment sums and kernels, Merkle trees, reward, etc.

pub fn verify_coinbase(&self) -> Result<(), Error>

Validate the coinbase.body.outputs generated by miners. Check the sum of coinbase-marked outputs match the sum of coinbase-marked kernels accounting for fees.