Struct epic_core::core::block::BlockHeader

pub struct BlockHeader {

    pub version: HeaderVersion,
    pub height: u64,
    pub prev_hash: Hash,
    pub prev_root: Hash,
    pub timestamp: DateTime<Utc>,
    pub output_root: Hash,
    pub range_proof_root: Hash,
    pub kernel_root: Hash,
    pub total_kernel_offset: BlindingFactor,
    pub output_mmr_size: u64,
    pub kernel_mmr_size: u64,
    pub pow: ProofOfWork,
    pub policy: u8,
    pub bottles: Policy,

}

Block header, fairly standard compared to other blockchains.

Fields
version: HeaderVersion

Version of the block

height: u64

Height of this block since the genesis block (height 0)

prev_hash: Hash

Hash of the block previous to this in the chain.

prev_root: Hash

Root hash of the header MMR at the previous header.

timestamp: DateTime<Utc>

Timestamp at which the block was built.

output_root: Hash

Merklish root of all the commitments in the TxHashSet

range_proof_root: Hash

Merklish root of all range proofs in the TxHashSet

kernel_root: Hash

Merklish root of all transaction kernels in the TxHashSet

total_kernel_offset: BlindingFactor

Total accumulated sum of kernel offsets since genesis block. We can derive the kernel offset sum for this block from the total kernel offset of the previous block header.

output_mmr_size: u64

Total size of the output MMR after applying this block

kernel_mmr_size: u64

Total size of the kernel MMR after applying this block

pow: ProofOfWork

Proof of work and related

policy: u8

The policy that generated this block

bottles: Policy

The bottles used in the feijoada algorithm

Implementations

impl BlockHeader

pub fn write_pre_pow<W: Writer>(&self, writer: &mut W) -> Result<(), Error>

Write the pre-hash portion of the header

pub fn pre_pow(&self) -> Vec<u8>

Return the pre-pow, unhashed Let the cuck(at)oo miner/verifier handle the hashing for consistency with how this call is performed everywhere else

pub fn pre_pow_hash(&self) -> [u8; 32]

pub fn total_difficulty(&self) -> Difficulty

Total difficulty accumulated by the proof of work on this header

pub fn overage(&self) -> i64

pub fn total_overage(&self, genesis_had_reward: bool) -> i64

The “total overage” to use when verifying the kernel sums for a full chain state. For a full chain state this is 0 - (height * reward).

pub fn total_kernel_offset(&self) -> BlindingFactor

Total kernel offset for the chain state up to and including this block.