Module epic_core::core::committed

The Committed trait and associated errors.

Enums

Error

Errors from summing and verifying kernel excesses via committed trait.

Traits

Committed

Implemented by types that hold inputs and outputs (and kernels) containing Pedersen commitments. Handles the collection of the commitments as well as their summing, taking potential explicit overages of fees into account.

Functions

sum_commits

Utility to sum positive and negative commitments, eliminating zero values

sum_kernel_offsets

Utility function to take sets of positive and negative kernel offsets as blinding factors, convert them to private key filtering zero values and summing all of them. Useful to build blocks.