Function epic_core::core::committed::sum_kernel_offsets

pub fn sum_kernel_offsets(
    positive: Vec<BlindingFactor>, 
    negative: Vec<BlindingFactor>
) -> Result<BlindingFactor, Error>

Utility function to take sets of positive and negative kernel offsets as blinding factors, convert them to private key filtering zero values and summing all of them. Useful to build blocks.