Struct epic_core::core::hash::HashWriter

pub struct HashWriter { /* fields omitted */ }

Serializer that outputs a hash of the serialized object

Implementations

impl HashWriter

pub fn finalize(self, output: &mut [u8])

Consume the HashWriter, outputting its current hash into a 32-byte array

pub fn into_hash(self) -> Hash

Consume the HashWriter, outputting a Hash corresponding to its current state