Trait epic_core::core::id::ShortIdentifiable

pub trait ShortIdentifiable {
    fn short_id(&self, hash: &Hash, nonce: u64) -> ShortId;
}

A trait for types that have a short_id (inputs/outputs/kernels)

Required methods

fn short_id(&self, hash: &Hash, nonce: u64) -> ShortId

The short_id of a kernel uses a hash built from the block_header and a connection specific nonce to minimize the effect of collisions.