Struct epic_core::core::transaction::TransactionBody

pub struct TransactionBody {
    pub inputs: Vec<Input>,
    pub outputs: Vec<Output>,
    pub kernels: Vec<TxKernel>,
}

TransactionBody is a common abstraction for transaction and block

Fields
inputs: Vec<Input>

List of inputs spent by the transaction.

outputs: Vec<Output>

List of outputs the transaction produces.

kernels: Vec<TxKernel>

List of kernels that make up this transaction (usually a single kernel).

Implementations

impl TransactionBody

pub fn empty() -> TransactionBody

Creates a new empty transaction (no inputs or outputs, zero fee).

pub fn sort(&mut self)

Sort the inputs|outputs|kernels.

pub fn init(
inputs: Vec<Input>,
outputs: Vec<Output>,
kernels: Vec<TxKernel>,
verify_sorted: bool
) -> Result<TransactionBody, Error>

Creates a new transaction body initialized with the provided inputs, outputs and kernels. Guarantees inputs, outputs, kernels are sorted lexicographically.

pub fn with_input(self, input: Input) -> TransactionBody

Builds a new body with the provided inputs added. Existing inputs, if any, are kept intact. Sort order is maintained.

pub fn with_output(self, output: Output) -> TransactionBody

Builds a new TransactionBody with the provided output added. Existing outputs, if any, are kept intact. Sort order is maintained.

pub fn with_kernel(self, kernel: TxKernel) -> TransactionBody

Builds a new TransactionBody with the provided kernel added. Existing kernels, if any, are kept intact. Sort order is maintained.

pub fn replace_kernel(self, kernel: TxKernel) -> TransactionBody

Builds a new TransactionBody replacing any existing kernels with the provided kernel.

pub fn fee(&self) -> u64

Total fee for a TransactionBody is the sum of fees of all fee carrying kernels.

pub fn body_weight(&self) -> usize

Calculate transaction weight

pub fn body_weight_as_block(&self) -> usize

Calculate weight of transaction using block weighing

pub fn weight(input_len: usize, output_len: usize, kernel_len: usize) -> usize

Calculate transaction weight from transaction details. This is non consensus critical and compared to block weight, incentivizes spending more outputs (to lower the fee).

pub fn weight_as_block(
input_len: usize,
output_len: usize,
kernel_len: usize
) -> usize

Calculate transaction weight using block weighing from transaction details. Consensus critical and uses consensus weight values.

pub fn lock_height(&self) -> u64

Lock height of a body is the max lock height of the kernels.

pub fn verify_features(&self) -> Result<(), Error>

Verify we have no invalid outputs or kernels in the transaction due to invalid features. Specifically, a transaction cannot contain a coinbase output or a coinbase kernel.

pub fn validate_read(&self, weighting: Weighting) -> Result<(), Error>

“Lightweight” validation that we can perform quickly during read/deserialization. Subset of full validation that skips expensive verification steps, specifically -

rangeproof verification

kernel signature verification

pub fn validate(&self, weighting: Weighting) -> Result<(), Error>

Validates all relevant parts of a transaction body. Checks the excess value against the signature as well as range proofs for each output.