Module epic_core::pow::cuckaroo

Implementation of Cuckaroo Cycle, based on Cuckoo Cycle designed by John Tromp. Ported to Rust from https://github.com/tromp/cuckoo.

Cuckaroo is an ASIC-Resistant variation of Cuckoo (CuckARoo) that’s aimed at making the lean mining mode of Cuckoo extremely ineffective. It is one of the 2 proof of works used in Epic (the other one being the more ASIC friendly Cuckatoo).

In Cuckaroo, edges are calculated by repeatedly hashing the seeds to obtain blocks of values. Nodes are then extracted from those edges.

Structs

CuckarooContext

Cuckatoo cycle context. Only includes the verifier for now.

Functions

new_cuckaroo_ctx

Instantiate a new CuckarooContext as a PowContext. Note that this can’t be moved in the PoWContext trait as this particular trait needs to be convertible to an object trait.