Trait epic_core::pow::PoWContext

pub trait PoWContext<T> where
    T: EdgeType, {
    fn set_header_nonce(
        &mut self, 
        header: Vec<u8>, 
        nonce: Option<u64>, 
        height: Option<u64>, 
        solve: bool
    ) -> Result<(), Error>;

    fn pow_solve(&mut self) -> Result<Vec<Proof>, Error>;

    fn verify(&mut self, proof: &Proof) -> Result<(), Error>;
}

Generic trait for a solver/verifier providing common interface into Cuckoo-family PoW Mostly used for verification, but also for test mining if necessary

Required methods

fn set_header_nonce(
&mut self,
header: Vec<u8>,
nonce: Option<u64>,
height: Option<u64>,
solve: bool
) -> Result<(), Error>

Sets the header along with an optional nonce at the end solve: whether to set up structures for a solve (true) or just validate (false)

fn pow_solve(&mut self) -> Result<Vec<Proof>, Error>

find solutions using the stored parameters and header

fn verify(&mut self, proof: &Proof) -> Result<(), Error>

Verify a solution with the stored parameters