Trait epic_core::ser::PMMRable

pub trait PMMRable: Writeable + Clone + Debug + DefaultHashable {
    type E: Readable + Writeable + Debug;
    fn as_elmt(&self) -> Self::E;

    fn elmt_size() -> Option<u16>;
}

Trait for types that can be added to a PMMR.

Associated Types

type E: Readable + Writeable + Debug

The type of element actually stored in the MMR data file. This allows us to store Hash elements in the header MMR for variable size BlockHeaders.

Required methods

fn as_elmt(&self) -> Self::E

Convert the pmmrable into the element to be stored in the MMR data file.

fn elmt_size() -> Option<u16>

Size of each element if “fixed” size. Elements are “variable” size if None.