Trait epic_core::ser::Readable

pub trait Readable where
    Self: Sized, {
    fn read(reader: &mut dyn Reader) -> Result<Self, Error>;
}

Trait that every type that can be deserialized from binary must implement. Reads directly to a Reader, a utility type thinly wrapping an underlying Read implementation.

Required methods

fn read(reader: &mut dyn Reader) -> Result<Self, Error>

Reads the data necessary to this Readable from the provided reader