Trait epic_core::ser::VerifySortedAndUnique

pub trait VerifySortedAndUnique<T> {
    fn verify_sorted_and_unique(&self) -> Result<(), Error>;
}

Collections of items must be sorted lexicographically and all unique.

Required methods

fn verify_sorted_and_unique(&self) -> Result<(), Error>

Verify a collection of items is sorted and all unique.