Trait epic_core::ser::Writeable

pub trait Writeable {
    fn write<W: Writer>(&self, writer: &mut W) -> Result<(), Error>;
}

Trait that every type that can be serialized as binary must implement. Writes directly to a Writer, a utility type thinly wrapping an underlying Write implementation.

Required methods

fn write<W: Writer>(&self, writer: &mut W) -> Result<(), Error>

Write the data held by this Writeable to the provided writer